% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_googleDrive.R
\name{auth_gdrive}
\alias{auth_gdrive}
\title{Authorize Google Drive access}
\usage{
auth_gdrive(config_path = "config.json", token_path = "tokens/gdrive_token",
  create_token = TRUE)
}
\arguments{
\item{config_path}{path where the config.json file is located}

\item{token_path}{path to the token file}

\item{create_token}{whether a new token should be created (mandatory, if an existing token is not available in the dir tokens/)}
}
\value{
A google drive token in the (newly created?) \code{tokens/gdrive_token} directory.
}
\description{
\code{auth_gdrive} generates a token which is saved to \code{tokens/gdrive_token} and used for accessing a google drive folder
}
\examples{
\dontrun{
auth_gdrive("config.json", create_token=TRUE)
}
}

