% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrTools_buildReport.R
\name{buildReport}
\alias{buildReport}
\title{Build an Rmarkdown report into a PDF}
\usage{
buildReport(rmarkdown_path, output_dir = getwd(), output_file, params)
}
\arguments{
\item{rmarkdown_path}{path of the rmarkdown file}

\item{output_dir}{directory the report should be written to}

\item{output_file}{name of the report file on output}

\item{params}{parameters to be passed on to the rmarkdown file}
}
\description{
Build an Rmarkdown report into a PDF
}

