% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config_template.R
\name{createConfigTemplate}
\alias{createConfigTemplate}
\title{Build config.json template}
\usage{
createConfigTemplate(config_path = "")
}
\arguments{
\item{config_path}{path where the config.json file is located}
}
\value{
null
}
\description{
\code{auth_gdrive} generates a template config file.
}
\examples{
\dontrun{
createConfigTemplate()
}
}

