% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_queries.R
\name{dbConnect_withConfig}
\alias{dbConnect_withConfig}
\title{Connect to DB using config.json file}
\usage{
dbConnect_withConfig(config_path)
}
\arguments{
\item{config_path}{Path to the config.json object.}
}
\value{
con Config object
}
\description{
Creates a connection object to a mysql DB based on a json configuration
}

