% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_googleDrive.R
\name{gdrive_auth}
\alias{gdrive_auth}
\title{Authorize Google Drive access}
\usage{
gdrive_auth(config_path = "config.json", create_token = TRUE)
}
\arguments{
\item{config_path}{path where the config.json file is located}

\item{create_token}{whether a new token should be created (mandatory, if an existing token is not available in the dir token/)}
}
\value{
A google drive token in the (newly created?) \code{token/gdrive_token} directory.
}
\description{
\code{gdrive_auth} generates a token which is saved to \code{token/gdrive_token} and used for accessing a google drive folder
}
\examples{
\dontrun{
gdrive_auth("config.json", create_token=TRUE)
}
}

