% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_googleDrive.R
\name{get_gdrive_queries}
\alias{get_gdrive_queries}
\title{Get Google Drive queries}
\usage{
get_gdrive_queries(config_path)
}
\arguments{
\item{config_path}{path where the config.json file is located}
}
\value{
list of \code{data.table}s which have the query information required to run google analytics queries
}
\description{
\code{get_gdrive_queries} returns a list of \code{data.table}s which have the query information required to run google analytics queries
}
\examples{
\dontrun{
get_gdrive_queries("config.json")
}
}

