% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_queries.R
\name{getDBdata}
\alias{getDBdata}
\title{Queries data base using MySQL config file & queries list object}
\usage{
getDBdata(queries, config_path = "config.json")
}
\description{
\code{getDBdata} Queries data base using MySQL config file & queries list object
}
\examples{
\dontrun{
queries <- getQueries()
getDBdata(queries)
}
}

