% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_queries.R
\name{getQueries}
\alias{getQueries}
\title{Creates a list object with labeled sql queries}
\usage{
getQueries(path = "", queries_folder = "queries", query_filter = "all",
  package = NULL)
}
\value{
list(Query1 = "SELECT * FROM abc")
}
\description{
\code{getQueries} Creates a list object with labeled sql queries
}
\examples{
\dontrun{
queries <- getQueries()
getDBdata(queries)
}
}

