% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_queries.R
\name{init_queries}
\alias{init_queries}
\title{Creates a queries folder as well as a template config.json doc}
\usage{
init_queries()
}
\description{
\code{init_queries} Sets up directory for reading sql queries out of a queries directory
}
\examples{
\dontrun{
init_queries()
}
}

