% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_jsonTools.R
\name{jsonFolderToDF}
\alias{jsonFolderToDF}
\title{Pulls data from folder of data.frame convertable json files and rbinds it into a data.table}
\usage{
jsonFolderToDF(path)
}
\value{
data.table
}
\description{
\code{init_queries} Pulls
}

