% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook.R
\name{query_facebook_pages}
\alias{query_facebook_pages}
\title{\code{query_facebook_pages} returns Facebook Page Data}
\usage{
query_facebook_pages(facebookPages, start_date, end_date,
  token_path = "tokens/fbInsights_token")
}
\arguments{
\item{facebookPages}{list of Facebook Page IDs}

\item{start_date}{start date as character string in \code{"\%Y\/\%m\/\%d"} format}

\item{end_date}{end date as character string in \code{"\%Y\/\%m\/\%d"} format}

\item{token}{path where the config.json file is located}
}
\value{
A data.table filled with great, interesting facebook page date!!
}
\description{
\code{query_facebook_pages} returns Facebook Page Data
}
\examples{
\dontrun{
query_facebook_pages("<Page/ID>", "2016/01/01", "2016/05/01")
}
}

