% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facebook.R
\name{query_facebook_posts}
\alias{query_facebook_posts}
\title{\code{query_facebook_posts} returns Facebook Post Data - query_facebook_pages needed to run before!}
\usage{
query_facebook_posts(token_path = "tokens/fbInsights_token",
  metrics = c("post_impressions_unique"))
}
\arguments{
\item{token}{path where the config.json file is located}
}
\value{
A data.table filled with great, interesting facebook post data!
}
\description{
\code{query_facebook_posts} returns Facebook Post Data - query_facebook_pages needed to run before!
}
\examples{
\dontrun{
query_facebook_posts(metrics = c("post_impressions_organic","post_impressions_paid"))
}
}

