% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_facebook.R
\name{query_facebook_single}
\alias{query_facebook_single}
\title{Fetch clean, streamlined Facebook Insights Data, from a data.table of queries}
\usage{
query_facebook_single(query, start_date, end_date, config_path, token_path,
  override_dates = override_dates)
}
\description{
Fetch clean, streamlined Facebook Insights Data, from a data.table of queries
}

