% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ga.R
\name{query_ga}
\alias{query_ga}
\title{Fetch clean, streamlined GoogleAnalytics Data, from a data.table of queries}
\usage{
query_ga(queries_table, start_date, end_date, config_path = "config.json",
  override_dates = TRUE, include_query_data = FALSE)
}
\arguments{
\item{queries_table}{data.table of queries}

\item{start_date}{start date as character string in \code{"\%Y-\%m-\%d"} format}

\item{end_date}{end date as character string in \code{"\%Y-\%m-\%d"} format}

\item{config_path}{path where the config.json file is located}

\item{override_dates}{whether all dates should be overriden with those in \code{start_date} & \code{end_date}}

\item{include_query_data}{whether the full query data should be included in a column as a json object}
}
\value{
A data.table filled with great, interesting googleAnalytics data!!
}
\description{
\code{query_ga} returns multiple GoogleAnalytics responses in a single data.table
}
\examples{
\dontrun{
query_ga(queries_table, "2015-01-01", "2015-02-01", config_path = "config.json", override_dates = TRUE)
}
}

