% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sc.R
\name{query_sc}
\alias{query_sc}
\title{Get Adobe Analytics Data}
\usage{
query_sc(queries_table, start_date, end_date, config_path = "config.json",
  override_dates = TRUE, include_query_data = FALSE)
}
\arguments{
\item{queries_table}{contains all queries}

\item{start_date}{contains the start date as date}

\item{end_date}{contains the end date as date}

\item{config_path}{contains the path of the config.json}

\item{override_dates}{set to true in order to override the query date}

\item{include_query_data}{set to true in order to include the query data to the output file}
}
\value{
RETURN data table with all queried data
}
\description{
Query Data from Adobe Analytics and get a data table as output
}
\examples{
# ADD EXAMPLES HERE
}

