% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_jsonTools.R
\name{saveRefreshData}
\alias{saveRefreshData}
\title{Pulls data from folder of data.frame convertable json files and rbinds it into a data.table}
\usage{
saveRefreshData(gatherDataFunct, label, refresh = FALSE)
}
\value{
data.table
}
\description{
\code{init_queries} Pulls
}

