% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzrtools_misc.R
\name{writeMarkdown}
\alias{writeMarkdown}
\title{Write Mardkown}
\usage{
writeMarkdown(input)
}
\arguments{
\item{input}{String}
}
\value{
RETURN mardown syntax with string
}
\description{
Get Markdown syntax to write text
}
\examples{
# writeMarkdown("# Headline")
}

