% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_facebook.R
\name{auth_fbInsights}
\alias{auth_fbInsights}
\title{Authorize Facebook Insights access}
\usage{
auth_fbInsights(config_path = "config.json",
  token_path = "tokens/fbInsights_token", create_token = TRUE)
}
\arguments{
\item{config_path}{path where the config.json file is located}

\item{token_path}{path to the token file}

\item{create_token}{whether a new token should be created (mandatory, if an existing token is not available in the dir tokens/)}
}
\value{
A google drive token in the (newly created?) \code{tokens/fbInsights_token} directory.
}
\description{
\code{auth_fbInsights} generates a token which is saved to \code{tokens/fbInsights_token} and used for accessing Facebook Insights
}
\examples{
\dontrun{
auth_fbInsights("config.json", create_token=TRUE)
}
}

